+++ test/test-weof.c
@@ -43,7 +43,7 @@ write_cb(evutil_socket_t fd, short event, void *arg)
 	const char *test = "test string";
 	int len;
 
-	len = send(fd, test, strlen(test) + 1, 0);
+	len = send(fd, test, (int)strlen(test) + 1, 0);
 
 	printf("%s: write %d%s\n", __func__,
 	    len, len ? "" : " - means EOF");
